Fixed some issues with other tests caused by updating the way Tone was being mocked.
